Updating the driver to allow read/writes to ioports. This is important to be able to access CSR registers that are on the board.  In the interest in keeping this driver as independent of the board as possible, I have removed specific references to the Concurrent in the driver.
Adding a header file for CSR registers in I/O memory
for the Concurrent SBC.  See that manual for more
information on which each of those registers do.  

The driver now checks the board to see if it can
determine the board type.  This is only possible
if the kernel has been compiled with the CONFIG_DMI
flag.  It has ioctl functions which gives the user
access to I/O ports, but only to those ports which
are specific to that board so as to not generate
a bad security risk.